Replace cookie provider: - change folder name. - split files to refactor code for appropriate files.
