+++ packages/extension/src/view/devtools/stateProviders/cookie/cookieProvider.tsx
@@ -16,7 +16,6 @@
 /**
  * External dependencies.
  */
-import { createContext } from 'use-context-selector';
 import React, {
   type PropsWithChildren,
   useEffect,
@@ -25,17 +24,11 @@ import React, {
   useRef,
   useMemo,
 } from 'react';
-import { noop } from '@ps-analysis-tool/design-system';
-import {
-  type TabCookies,
-  type TabFrames,
-  UNKNOWN_FRAME_KEY,
-} from '@ps-analysis-tool/common';
+import { type TabCookies, UNKNOWN_FRAME_KEY } from '@ps-analysis-tool/common';
 
 /**
  * Internal dependencies.
  */
-import useContextSelector from '../../../../utils/useContextSelector';
 import {
   ALLOWED_NUMBER_OF_TABS,
   DEVTOOLS_CLOSE,
@@ -48,60 +41,9 @@ import setDocumentCookies from '../../../../utils/setDocumentCookies';
 import { useSettingsStore } from '../syncSettingsStore';
 import { getTab } from '../../../../utils/getTab';
 import getFramesForCurrentTab from '../../../../utils/getFramesForCurrentTab';
+import Context, { type CookieStoreContext } from './context';
 
-export interface CookieStoreContext {
-  state: {
-    tabCookies: TabCookies | null;
-    tabUrl: string | null;
-    loading: boolean;
-    tabFrames: TabFrames | null;
-    selectedFrame: string | null;
-    returningToSingleTab: boolean;
-    isCurrentTabBeingListenedTo: boolean;
-    isInspecting: boolean;
-    contextInvalidated: boolean;
-    canStartInspecting: boolean;
-    tabToRead: string | null;
-    frameHasCookies: Record<string, boolean>;
-  };
-  actions: {
-    setSelectedFrame: (key: string | null) => void;
-    setIsInspecting: React.Dispatch<React.SetStateAction<boolean>>;
-    changeListeningToThisTab: () => void;
-    getCookiesSetByJavascript: () => void;
-    setContextInvalidated: React.Dispatch<React.SetStateAction<boolean>>;
-    setCanStartInspecting: React.Dispatch<React.SetStateAction<boolean>>;
-  };
-}
-
-const initialState: CookieStoreContext = {
-  state: {
-    tabCookies: null,
-    tabUrl: null,
-    tabFrames: null,
-    selectedFrame: null,
-    loading: true,
-    isCurrentTabBeingListenedTo: false,
-    returningToSingleTab: false,
-    isInspecting: false,
-    contextInvalidated: false,
-    canStartInspecting: false,
-    tabToRead: null,
-    frameHasCookies: {},
-  },
-  actions: {
-    setSelectedFrame: noop,
-    changeListeningToThisTab: noop,
-    setIsInspecting: noop,
-    getCookiesSetByJavascript: noop,
-    setContextInvalidated: noop,
-    setCanStartInspecting: noop,
-  },
-};
-
-export const Context = createContext<CookieStoreContext>(initialState);
-
-export const Provider = ({ children }: PropsWithChildren) => {
+const Provider = ({ children }: PropsWithChildren) => {
   const [loading, setLoading] = useState<boolean>(true);
   const loadingTimeout = useRef<ReturnType<typeof setTimeout> | null>(null);
   const [tabToRead, setTabToRead] = useState<string | null>(null);
@@ -450,19 +392,4 @@ export const Provider = ({ children }: PropsWithChildren) => {
   );
 };
 
-export function useCookieStore(): CookieStoreContext;
-export function useCookieStore<T>(
-  selector: (state: CookieStoreContext) => T
-): T;
-
-/**
- * Cookie store hook.
- * @param selector Selector function to partially select state.
- * @returns selected part of the state
- */
-export function useCookieStore<T>(
-  selector: (state: CookieStoreContext) => T | CookieStoreContext = (state) =>
-    state
-) {
-  return useContextSelector(Context, selector);
-}
+export default Provider;
