+++ packages/extension/src/view/devtools/stateProviders/cookie/useCookie.ts
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2023 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/**
+ * Internal dependencies.
+ */
+import useContextSelector from '../../../../utils/useContextSelector';
+import Context, { type CookieStoreContext } from './context';
+
+export function useCookie(): CookieStoreContext;
+export function useCookie<T>(selector: (state: CookieStoreContext) => T): T;
+
+/**
+ * Cookie store hook.
+ * @param selector Selector function to partially select state.
+ * @returns selected part of the state
+ */
+export function useCookie<T>(
+  selector: (state: CookieStoreContext) => T | CookieStoreContext = (state) =>
+    state
+) {
+  return useContextSelector(Context, selector);
+}
